Code refactor and response improvements
