+++ examples/ssx-test-serverless-dynamodb-api/src/functions/ssx-login/index.ts
@@ -1,5 +1,5 @@
-import { signInSchema } from './schema';
 import { handlerPath } from '@libs/handler-resolver';
+import { signInSchema } from '@libs/schemas';
 
 export default {
   handler: `${handlerPath(__dirname)}/handler.main`,
