--- examples/ssx-test-serverless-dynamodb-api/src/functions/ssx-me/schema.ts
@@ -1,7 +1,0 @@
-export const getMeSchema = {
-  type: "object",
-  properties: {
-    address: { type: 'string' },
-  },
-  required: ['address']
-} as const;
