+++ examples/ssx-test-serverless-dynamodb-api/src/functions/ssx-nonce/index.ts
@@ -1,5 +1,5 @@
-import { getNonceSchema } from './schema';
 import { handlerPath } from '@libs/handler-resolver';
+import { requireAddressSchema } from '@libs/schemas';
 
 export default {
   handler: `${handlerPath(__dirname)}/handler.main`,
@@ -10,7 +10,7 @@ export default {
         path: 'ssx-nonce',
         request: {
           schemas: {
-            'application/json': getNonceSchema,
+            'application/json': requireAddressSchema,
           },
         },
       },
