--- examples/ssx-test-serverless-dynamodb-api/src/functions/ssx-nonce/schema.ts
@@ -1,7 +1,0 @@
-export const getNonceSchema = {
-  type: "object",
-  properties: {
-    address: { type: 'string' }
-  },
-  required: ['address']
-} as const;
