+++ examples/ssx-test-serverless-dynamodb-api/src/libs/api-gateway.ts
@@ -4,9 +4,9 @@ import type { FromSchema } from "json-schema-to-ts";
 type ValidatedAPIGatewayProxyEvent<S> = Omit<APIGatewayProxyEvent, 'body'> & { body: FromSchema<S> }
 export type ValidatedEventAPIGatewayProxyEvent<S> = Handler<ValidatedAPIGatewayProxyEvent<S>, APIGatewayProxyResult>
 
-export const formatJSONResponse = (response: Record<string, unknown>) => {
-  return {
-    statusCode: 200,
-    body: JSON.stringify(response)
-  }
+export const formatJSONResponse = (statusCode: number, response: Record<string, unknown>) => {
+    return {
+        statusCode,
+        body: JSON.stringify(response)
+    }
 }
